refactor: move AutoinstallDirectModel into separate file
